test: reuse shared metrics recorder
